add xmls to maven classpath.  change Double to BigDecimal to have precise values2

+++ fund-splitting-checker/src/main/java/com/silvanh/xmlModels/PositionModel.java
@@ -1,17 +1,19 @@
 package com.silvanh.xmlModels;
 
 
+import java.math.BigDecimal;
+
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty;
 
 public class PositionModel {
 
     @JacksonXmlProperty(localName = "percentage")
-    private Double percentage;
+    private BigDecimal percentage;
 
     @JacksonXmlProperty(isAttribute = true)
     private String id; 
 
-    public Double getPercentage() {
+    public BigDecimal getPercentage() {
         return percentage;
     }
